# openapi.yaml
openapi: 3.0.3
info:
  title: FindWashroom Backend API
  description: >-
    Complete RESTful API for FindWashroom - a property/washroom booking and management platform.
    Supports multi-role authentication (admin, owner, customer, cleaner), property management,
    booking system, payments (Razorpay), reviews, complaints, and wallet management.
  version: 1.0.0
  contact:
    name: FindWashroom Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.findwashroom.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: Sanctum API token. Get token from login endpoints.
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
          additionalProperties: true
        meta:
          type: object
          nullable: true
          additionalProperties: true
      required:
        - success
        - message
...
    WalletTransaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        wallet_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
          format: float
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
paths:
...
  /v1/service-units/{id}:
    get:
      tags: [Service Units]
      summary: Get Service Unit Details
      operationId: getServiceUnitDetails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Service unit details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /v1/products/{id}:
    get:
      tags: [Products]
      summary: Get Product Details
      operationId: getProductDetails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /v1/wallet/transactions:
    get:
      tags: [Wallet]
      summary: Get Wallet Transactions
      operationId: getWalletTransactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transaction history
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/WalletTransaction'
...
  /v1/wallet/request-payout:
    post:
      tags: [Wallet]
      summary: Request Payout
      operationId: requestPayout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  format: float
                  minimum: 1
      responses:
        '200':
          description: Payout requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Trust
...